+++ imgdoc2API/imgdoc2statisticsinterop.h
